move watertransp into watertransp dir:
Moved watertransp into the watertransp dir.
All of its routines were called in runspace.
