+++ CMakeLists.txt
@@ -302,6 +302,7 @@ ELSEIF("${COMPILE_TARGET}" STREQUAL "summa_openwq")
                         ${OPENWQ_DIR}/src/extwatflux_ss/*.cpp
                         ${OPENWQ_DIR}/src/initiate/*.cpp
                         ${OPENWQ_DIR}/src/chem/*.cpp
+                        ${OPENWQ_DIR}/src/watertransp/*.cpp
                         ${PARENT_DIR}/OpenWQ_hydrolink.cpp
                         ${PARENT_DIR}/OpenWQ_interface.cpp)
