+++ src/OpenWQ_readjson.h
@@ -27,7 +27,7 @@
 #include <unistd.h>
 #include <vector>
 
-#include "exprtk.hpp"
+#include "global/exprtk.hpp"
 
 #include <cstdio>
