+++ src/chem/OpenWQ_chem.hpp
@@ -22,7 +22,7 @@
 #include <armadillo>
 #include <string>
 #include <algorithm>
-#include "exprtk.hpp"
+#include "global/exprtk.hpp"
 #include <cstdio>
 #include <tuple>
 #include <vector>
